Updated the prison scene
Removed a lot of the assets and resized them as I saw fit. Added a custom prison prefab to the prison scene from another scene I made. Will be working on the prison itself more, it is too small and cramp as of right now. Will be adding more rooms when the time comes. For now, I will need to implement the most important functionality, the VR and the voice recognition implementation for commands like, "Return Home" or "Light On" and such.