Some Tests
- added a commented out segment of  code to the simple collision damage script (found this to be the spot to edit)
- add an additional npc while testing
- added the infinite ammo gun to the bottom inventory slot for testing
- tested the ability to put ragdoll effects on npc (could not get to work)
Note: I believe I am close to the solution but running into problems due to gaps c knowledge. My idea was to call the deathController script to the Simple Collision Script (after realizing the projectile script did not house what I was looking for) and use the script to pass functionality of the death Controller when the player shoots an npc under the tag guy. In psuedo code terms, if there is a guy (guy =! null) when hit then destroy the object of guy.